Change uint256 to shorter type

+++ vyper-immutable-test/contracts/ImmutableTest.vy
@@ -1,10 +1,10 @@
 event Num:
-    num: uint256
+    num: int128
 
-imm: immutable(uint256)
+imm: immutable(int128)
 
 @external
-def __init__(x: uint256):
+def __init__(x: int128):
     imm = x
 
 @external
